fix: fix remove and keyed entries
